\name{DPMMmlogit-package}
\alias{DPMMmlogit-package}
\alias{DPMMmlogit}
\docType{package}
\title{DPMMmlogit
\packageTitle{DPMMmlogit}
}
\description{
\packageDescription{DPMMmlogit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DPMMmlogit}
\packageIndices{DPMMmlogit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DPMMmlogit}

Maintainer: \packageMaintainer{DPMMmlogit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
